requirements.txt
langchain-0.1.17
langchain-text-splitters-0.0.1
langchain_openai-0.1.4
openai-1.25.0
python-dotenv-1.0.1
streamlit-1.33.0